+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,38 +1,22 @@
 package roomescape.controller;
 
-import java.net.URI;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.CreateReservationTimeResponse;
-import roomescape.dto.ReservationTimeRequest;
 import roomescape.dto.ReservationTimeResponse;
 import roomescape.service.ReservationTimeService;
 
 @RestController
 @RequestMapping("/times")
 public class ReservationTimeController {
-    private static final String LOCATION_DEFAULT_VALUE = "/times/";
-
     private final ReservationTimeService reservationTimeService;
 
     public ReservationTimeController(ReservationTimeService reservationTimeService) {
         this.reservationTimeService = reservationTimeService;
-    }
-
-    @PostMapping
-    public ResponseEntity<CreateReservationTimeResponse> addReservationTime(@RequestBody ReservationTimeRequest request) {
-        CreateReservationTimeResponse response = reservationTimeService.addReservationTime(request);
-        return ResponseEntity.created(URI.create(LOCATION_DEFAULT_VALUE + response.id()))
-                .body(response);
     }
 
     @GetMapping
@@ -40,12 +24,5 @@ public class ReservationTimeController {
                                                                              @RequestParam(value = "date") LocalDate date) {
         List<ReservationTimeResponse> responses = reservationTimeService.getReservationTimes(themeId, date);
         return ResponseEntity.ok(responses);
-    }
-
-    @DeleteMapping("/{timeId}")
-    public ResponseEntity<Void> deleteReservationTime(@PathVariable("timeId") long reservationTimeId) {
-        reservationTimeService.deleteReservationTime(reservationTimeId);
-        return ResponseEntity.noContent()
-                .build();
     }
 }
